Add settings to hide blocks (#5306)

+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/BlockCollection/SortableBlockList.js
@@ -12,9 +12,11 @@ type Props = {|
     disabled: boolean,
     expandedBlocks: Array<boolean>,
     generatedBlockIds: Array<number>,
+    icons?: Array<Array<string>>,
     onCollapse: (index: number) => void,
     onExpand: (index: number) => void,
-    onRemove: (index: number) => void,
+    onRemove?: (index: number) => void,
+    onSettingsClick?: (index: number) => void,
     onTypeChange?: (type: string | number, index: number) => void,
     renderBlockContent: RenderBlockContentCallback,
     types?: {[key: string]: string},
@@ -39,7 +41,18 @@ class SortableBlockList extends React.Component<Props> {
 
     handleRemove = (index: number) => {
         const {onRemove} = this.props;
-        onRemove(index);
+
+        if (onRemove) {
+            onRemove(index);
+        }
+    };
+
+    handleSettingsClick = (index: number) => {
+        const {onSettingsClick} = this.props;
+
+        if (onSettingsClick) {
+            onSettingsClick(index);
+        }
     };
 
     handleTypeChange = (type: string | number, index: number) => {
@@ -51,7 +64,17 @@ class SortableBlockList extends React.Component<Props> {
     };
 
     render() {
-        const {disabled, expandedBlocks, generatedBlockIds, onRemove, renderBlockContent, types, value} = this.props;
+        const {
+            disabled,
+            expandedBlocks,
+            generatedBlockIds,
+            icons,
+            onRemove,
+            onSettingsClick,
+            renderBlockContent,
+            types,
+            value,
+        } = this.props;
 
         const sortableBlockListClass = classNames(
             sortableBlockListStyles.sortableBlockList,
@@ -66,11 +89,13 @@ class SortableBlockList extends React.Component<Props> {
                     <SortableBlock
                         activeType={block.type}
                         expanded={!disabled && expandedBlocks[index]}
+                        icons={icons && icons[index]}
                         index={index}
                         key={generatedBlockIds[index]}
                         onCollapse={this.handleCollapse}
                         onExpand={this.handleExpand}
                         onRemove={onRemove ? this.handleRemove : undefined}
+                        onSettingsClick={onSettingsClick ? this.handleSettingsClick : undefined}
                         onTypeChange={this.handleTypeChange}
                         renderBlockContent={renderBlockContent}
                         sortIndex={index}
